Submissão 5 do mooshak

+++ Funcoes_aux.c
@@ -165,10 +165,10 @@ void aresta_barata(int **maze, int lin, int col, int vertices, G *g)
 {
     int k = 0, a = 0, b = 0;
     ladj *aux;
-    short p;
+    int p;
     for (p = 0; p <= lin; p++)
     {
-        for (short q = 0; q <= col; q++)
+        for (int q = 0; q <= col; q++)
         {
             if ((k = FA5(maze, p, q, lin, col)) > 0) /*1- parte horizontalmente; 2- parte verticalmente; 3- parte nas duas direções*/
             {
